pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cwl-loader"
dynamic = ["version"]
description = 'Utility to load a CWL (remote) document'

readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Fabrice Brito", email = "info@terradue.com" },
  { name = "Simone Tripodi", email = "info@terradue.com"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "cwl-utils>=0.40",
  "cwltool>=3.1.20250110105449",
  "ruamel.yaml>=0.18.15",
  "loguru"
]

[project.urls]
Documentation = "https://github.com/eoap/cwl-loader#readme"
Issues = "https://github.com/eoap/cwl-loader/issues"
Source = "https://github.com/eoap/cwl-loader"

[tool.hatch.version]
path = "src/cwl_loader/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/cwl_loader tests}"

[tool.coverage.run]
source_pkgs = ["cwl_loader", "tests"]
branch = true
parallel = true
omit = [
  "src/cwl_loader/__about__.py",
]

[tool.coverage.paths]
cwl_loader = ["src/cwl_loader", "*/cwl_loader/src/cwl_loader"]
tests = ["tests", "*/cwl_loader/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]


[tool.hatch.envs.test]
skip-install = false

dependencies = [
  "cwl-utils==0.37",
  "loguru",
  "pyyaml",
  "nose2",
]


[tool.hatch.envs.test.scripts]
test = "hatch run nose2"
testv = "hatch run nose2 --verbose"
cov = ["coverage run --source=src/cwl_loader -m nose2", "coverage report"]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12", "3.13"]
